Extract acceptors from sockets in both epoll and select
