+++ src/corosio/src/select_context.cpp
@@ -14,6 +14,7 @@
 
 #include "src/detail/select/scheduler.hpp"
 #include "src/detail/select/sockets.hpp"
+#include "src/detail/select/acceptors.hpp"
 
 #include <thread>
